+++ src/wnm/__main__.py
@@ -23,6 +23,7 @@ from wnm.models import Machine, Node
 from wnm.utils import (
     get_antnode_version,
     get_machine_metrics,
+    get_system_start_time,
     update_counters,
 )